+++ app/src/main/java/com/arstkn/lock_screen_activity/App.kt
@@ -20,7 +20,7 @@ class App : Application() {
             val channel = NotificationChannel(
                 CHANNEL_ID,
                 "Example Channel",
-                NotificationManager.IMPORTANCE_HIGH
+                NotificationManager.IMPORTANCE_DEFAULT
             )
             val manager =
                 getSystemService(NotificationManager::class.java)
