+++ app/src/main/java/com/arstkn/lock_screen_activity/KeepScreenActivity.kt
@@ -1,8 +1,10 @@
 package com.arstkn.lock_screen_activity
 
 import android.app.NotificationManager
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import android.view.WindowManager
 import androidx.core.app.NotificationManagerCompat
 import kotlinx.android.synthetic.main.activity_keep_screen.*
@@ -12,12 +14,19 @@ class KeepScreenActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        Log.d("KeepScreenActivity", "onCreate start")
+
         setContentView(R.layout.activity_keep_screen)
         notificationManager = NotificationManagerCompat.from(this)
 
+        val service = Intent(this, NotificationService::class.java)
+        startForegroundService(service)
+
         stop_activity.setOnClickListener {
-            finish()
-            notificationManager!!.cancel(1)
+            stopService(service)
+
+            finishAndRemoveTask()
         }
 
         window.addFlags(
