+++ app/src/main/java/com/arstkn/lock_screen_activity/NotificationService.kt
@@ -1,0 +1,33 @@
+package com.arstkn.lock_screen_activity
+
+import android.app.PendingIntent
+import android.app.Service
+import android.content.Intent
+import android.os.IBinder
+import android.util.Log
+import androidx.core.app.NotificationCompat
+
+class NotificationService : Service() {
+
+    override fun onCreate() {
+        super.onCreate()
+
+        val intent = Intent(this, KeepScreenActivity::class.java)
+        val pendingIntent = PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT)
+
+        val notification = NotificationCompat.Builder(this, App.CHANNEL_ID)
+            .setSmallIcon(android.R.drawable.ic_menu_call)
+            .setContentTitle("start KeepScreenActivity!!")
+            .setContentText("content text")
+            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+            .setFullScreenIntent(pendingIntent, true)
+            .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
+            .setWhen(System.currentTimeMillis())
+
+        startForeground(1, notification.build())
+    }
+
+    override fun onBind(intent: Intent): IBinder {
+        TODO("Return the communication channel to the service.")
+    }
+}
